requirements.txt
fastapi~=0.104.1
uvicorn~=0.24.0
python-dotenv~=1.0.0
supabase~=2.0.3
pandas~=2.1.3
matplotlib~=3.8.1